Added a configuration file to locate the executable.
Added configuration file mopac.ini and mopac_step.ini to control and
tune the use of MOPAC and this stage. Add ConfigArgParse to do this.

Also automagically add THREADS=n keyword and MKL_NUM_THREADS
environment variable to provide finer control over the run.

And fixed a bug parsing the AUX file when the moments of intertia run
together on a line due to FORTRAN formatting limitations.