Changed to externalize the traversal of the grsph from the nodes.
Originally the nodes had a next function that returned the next node
in traversal of the graph. This required the nodes to reference the
graph itself, and is part fo the reason the  tk nodes also contain the
nodes. All of this leads to mess encapsulation.

These changes move the traversal of the graph outside of the
nodes. Now the nodes return the tag of the edge to traverse and the
external code uses this information to find the next node in the
graph.

By default the tag of the next edge is 'next'. For simple nodes this
is all that is needed. More complex nodes such as loops need to
override the method as appropriate.