Cleaning up code structure and dependencies.
Removed references to flowcharts in the nodes, and added in the edges
to/from the node. Working towards storing the flowchart directly from
the graphics nodes and being able to read back in either as graphics
nodes or computational nodes. To simplify this, I have added an abstract base class
NodeBase that both nongraphical and graphical nodes will use, and have
now done the same with flowcharts.

Starting to work on the MOPAC plug-in.

The final goal is to be able to run the non-graphical nodes methods
such as run in separate environments.